.DollarNames.GlobalOptionsFun {GlobalOptions}	R Documentation

The .DollarNames method for the GlobalOptionsFun class

Description

The .DollarNames method for the GlobalOptionsFun class

Usage

## S3 method for class 'GlobalOptionsFun'
.DollarNames(x, pattern = "")

Arguments

x
the object returned by set_opt or setGlobalOptions.

pattern
pattern, please ignore it.

Details

This makes the option object looks like a list that it allows option name completion after $.

Author(s)

z.gu@dkfz.de

Examples

# There is no example
NULL

[Package GlobalOptions version 0.1.2 Index]