set_opt {GlobalOptions}	R Documentation

Produce a function which can get or set global options

Description

Produce a function which can get or set global options

Usage

set_opt(...)

Arguments

...
all go to setGlobalOptions

Details

This is just a short name for setGlobalOptions.

Author(s)

z.gu@dkfz.de

Examples

# There is no example
NULL

[Package GlobalOptions version 0.1.2 Index]